Marble 1710 save supplemental data records for items for image/media (#312)
* MARBLE 1710 Save SupplementalData Records

* Fixed a bug in Curate export to save pdfs as media records instead of image records.
* Added folder for manage_storage_gateway_files to hold code for copying from the RBSC bucket to the libnd-smb-marble storage gateway, including rebuilding the storage gateway cache.
* Added code to clean_up_dynamo to insert SupplementalData records for Aleph and ArchivesSpace items to hold image and media metadata.  This is a prerequisite to stopping harvesting image references from those source systems.
* Added ability to set a parameter to recreate standard json locally for curate export.
* Changed standard_json_helpers.py to add itemGroupId, mediaGroupId, and objectFileGroupId to file-level records as appropriate.
* Code now exists to save SupplementalData records.  This will need to be re-run after harvests are re-run.
* dynamo_helpers wasn't adding mediaServer and mediaResourceId for pdfs.  This is now fixed.
* added clean_up_dynamo/find_pdf_file_records.py to generate a list of pdf files to copy.  This generates 2 files, copy_pdf_files.sh (which I executed to copy pdf files to libnd-smb-marble) and creates a csv file in case we need to reference the objectFileGroupId of these files in the future.

+++ clean_up_dynamo/handler.py
@@ -1,12 +1,14 @@
 """ handler """
 
+import _set_path  # noqa
 import boto3
-from boto3.dynamodb.conditions import Key  # , Attr
+from boto3.dynamodb.conditions import Key, Attr
 from botocore.exceptions import ClientError
 import os
 from pathlib import Path
 import sentry_sdk   # noqa: E402
 from sentry_sdk.integrations.aws_lambda import AwsLambdaIntegration
+from dynamo_helpers import add_supplemental_data_keys
 
 
 if 'SENTRY_DSN' in os.environ:
@@ -14,9 +16,15 @@ if 'SENTRY_DSN' in os.environ:
 
 
 def run(event: dict, context: dict) -> dict:
-    event['fileRecordsDeleted'] = delete_file_records(event.get('website-metadata-tablename'), event.get('deleteAllFileRecords', False))  # 2252 .xml files deleted
-    event['fileToHarvestRecordsDeleted'] = delete_file_to_process_records(event.get('website-metadata-tablename'), event.get('deleteAllFileToProcessRecords', False))  # 5488 records deleted
-    delete_specific_file_records(event.get('website-metadata-tablename'))
+    if event.get('deleteAllFileRecords', False):
+        event['fileRecordsDeleted'] = delete_file_records(event.get('website-metadata-tablename'), event.get('deleteAllFileRecords', False))  # 2252 .xml files deleted
+    if event.get('deleteAllFileToProcessRecords', False):
+        event['fileToHarvestRecordsDeleted'] = delete_file_to_process_records(event.get('website-metadata-tablename'), event.get('deleteAllFileToProcessRecords', False))  # 5488 records deleted
+    if event.get('deleteSpecificFileRecords', False):
+        delete_specific_file_records(event.get('website-metadata-tablename'))
+    if event.get('insertSupplementalDataRecords', False):
+        event['alephSupplementalDataRecordsInserted'] = insert_supplemental_data_records(event.get('website-metadata-tablename'), 'Aleph')
+        event['archivesSpaceSupplementalDataRecordsInserted'] = insert_supplemental_data_records(event.get('website-metadata-tablename'), 'ArchivesSpace')
     return event
 
 
@@ -152,6 +160,90 @@ def delete_file_to_process_records(table_name: str, delete_all_records: bool):
     return records_deleted
 
 
+def insert_supplemental_data_records(table_name: str, source_system: str):
+    """ Insert SupplementalData records for a given source_system to include Item fields: id, objectFileGroupId, imageGroupId, mediaGroupId, defaultFilePath """
+    print("inserting SupplementalData records into ", table_name, 'for', source_system)
+    kwargs = {}
+    kwargs['FilterExpression'] = Attr("TYPE").eq("Item") and Attr("sourceSystem").eq(source_system)
+    kwargs['ProjectionExpression'] = 'id, objectFileGroupId, imageGroupId, mediaGroupId, defaultFilePath'
+    done = False
+    records_inserted = 0
+    table = boto3.resource('dynamodb').Table(table_name)
+    while not done:
+        results = scan_dynamo_records(table_name, **kwargs)
+        # Note that batch.put_item overwrites any potentially already existing SupplementalData records.  We can't do this, so we must use update_item.  But I'm keeping this here for future reference
+        # with table.batch_writer() as batch:
+        for record in results.get('Items', []):
+            if record.get('objectFileGroupId') or record.get('imageGroupId') or record.get('mediaGroupId') or record.get('defaultFilePath'):
+                working_set_json = add_supplemental_data_keys(record.copy())
+                update_expression = 'SET '
+                expression_attribute_names = {}
+                expression_attribute_values = {}
+                for k, v in working_set_json.items():
+                    if k not in ('PK', 'SK'):  # Note:  we can't update any part of the primary key
+                        update_expression += '#' + k + ' = :' + k + ', '
+                        expression_attribute_names['#' + k] = k
+                        expression_attribute_values[':' + k] = v
+                update_expression = update_expression[:-2]  # remove the trailing comma and space from the update expression string
+                response = table.update_item(
+                    Key={'PK': working_set_json.get('PK'), 'SK': working_set_json.get('SK')},
+                    UpdateExpression=update_expression,
+                    ExpressionAttributeNames=expression_attribute_names,
+                    ExpressionAttributeValues=expression_attribute_values,
+                    ReturnValues="ALL_NEW"
+                )
+                if response.get('ResponseMetadata').get('HTTPStatusCode') != 200:
+                    print("response = ", response)
+                    print(1 / 0)
+                # supplemental_data_record = add_supplemental_data_keys(record.copy())
+                # print("supplemental_data_record = ", supplemental_data_record)
+                # Note that batch.put_item overwrites any potentially already existing SupplementalData records.  But I'm keeping this here for future reference.
+                # batch.put_item(supplemental_data_record)
+                records_inserted += 1
+        if results.get('LastEvaluatedKey'):
+            kwargs['ExclusiveStartKey'] = results.get('LastEvaluatedKey')
+        else:
+            done = True
+    return records_inserted
+
+
+def find_item_records_with_images(table_name: str):
+    kwargs = {}
+    kwargs['FilterExpression'] = Attr("TYPE").eq("Item") and Attr("imageGroupId").exists()
+    kwargs['ProjectionExpression'] = 'id, imageGroupId'
+    continued_count = 0
+    record_count = 0
+    unique_imageGroupIds_referenced = []
+    while True:
+        results = scan_dynamo_records(table_name, **kwargs)
+        for record in results.get('Items', []):
+            record_count += 1
+            image_group_id = record.get('imageGroupId')
+            if image_group_id:
+                record_count += 1
+                if image_group_id not in unique_imageGroupIds_referenced:
+                    unique_imageGroupIds_referenced.append(image_group_id)
+
+        if results.get('LastEvaluatedKey'):
+            kwargs['ExclusiveStartKey'] = results.get('LastEvaluatedKey')
+            continued_count += 1
+        else:
+            break
+    print("continued_cound =", continued_count, "record_count =", record_count, len(unique_imageGroupIds_referenced))
+    return unique_imageGroupIds_referenced
+
+
+def scan_dynamo_records(table_name: str, **kwargs) -> dict:
+    """ very generic dynamo scan """
+    response = {}
+    try:
+        table = boto3.resource('dynamodb').Table(table_name)
+        response = table.scan(**kwargs)
+    except ClientError as ce:
+        sentry_sdk.capture_exception(ce)
+    return response
+
+
 def query_dynamo_records(table_name: str, **kwargs) -> dict:
     """ very generic dynamo query """
     response = {}
@@ -175,7 +267,6 @@ def delete_dynamo_record(table_name: str, pk: str, sk: str) -> dict:
     return response
 
 # setup:
-# export SSM_KEY_BASE=/all/stacks/steve-manifest
 # aws-vault exec testlibnd-superAdmin
 # python -c 'from handler import *; test()'
 
@@ -185,12 +276,35 @@ def delete_dynamo_record(table_name: str, pk: str, sk: str) -> dict:
 
 def test(identifier=""):
     """ test exection """
+    # print(find_item_records_with_images('steve-manifest-websiteMetadata470E321C-1D6R3LX7EI284'))
     event = {}
-    event['website-metadata-tablename'] = 'steve-manifest-websiteMetadata470E321C-1NPIJYOXUCVHU'
-    event['website-metadata-tablename'] = 'marbleb-prod-manifest-websiteMetadata470E321C-5EJSG31E16Z7'
+    testlibnd_tables = [
+        'steve-manifest-websiteMetadata470E321C-1D6R3LX7EI284',
+        'jon-test-manifest-websiteMetadata470E321C-ZCSU70JC12I0',
+        'jon-prod-manifest-websiteMetadata470E321C-8NG755QB2S5I',
+        'mlk-manifest-websiteMetadata470E321C-EU6Z5BXP1WVB',
+        'sm-test-manifest-websiteMetadata470E321C-E5FXU8HUIWQG',
+        'sm-prod-manifest-websiteMetadata470E321C-HO7FZQXZXI8M',
+        'testlib-prod-manifest-websiteMetadata470E321C-1XA9OOG7PJWEE',
+        'testlib-test-manifest-websiteMetadata470E321C-1XQ2EFEWM3UXZ'
+    ]
+    libnd_tables = [
+        'marbleb-prod-manifest-websiteMetadata470E321C-5EJSG31E16Z7',
+        'marbleb-test-manifest-websiteMetadata470E321C-JJG277N1OMMC'
+    ]
+    tables_to_process = testlibnd_tables
+    tables_to_process = libnd_tables
+
+    # event['website-metadata-tablename'] = 'steve-manifest-websiteMetadata470E321C-1D6R3LX7EI284'
+    # event['website-metadata-tablename'] = 'marbleb-prod-manifest-websiteMetadata470E321C-5EJSG31E16Z7'
     # event['website-metadata-tablename'] = 'marbleb-test-manifest-websiteMetadata470E321C-JJG277N1OMMC'
     event['deleteAllFileRecords'] = False
     event['deleteAllFileToProcessRecords'] = False
-    # print(1 / 0)  # make sure we don't accidentally run this
-    event = run(event, {})
-    print(event)
+    event['deleteSpecificFileRecords'] = False
+    event['insertSupplementalDataRecords'] = False
+
+    # Do processing desired for all tables listed
+    for table_name in tables_to_process:
+        event['website-metadata-tablename'] = table_name
+        event = run(event, {})
+        print(table_name, event)
